pyproject.toml
[tool.poetry]
name = "telegram-notifier"
version = "0.1.0"
description = "Notify you when something is mentioned in a telegram channel"
authors = ["Pablo Cabezas <headsrooms@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
telethon = "^1.18.2"
12factor-configclasses = {extras = ["env"], version = "^0.3.0"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
